Introduce the `rules lint` command
This command lints a rules file. The linter's aim is not verify correctness but just YAML and PromQL expression formatting within the rule file. It is meant to always edit in place but has a flag to perform a dry-run.

This closes https://github.com/grafana/cortex-tools/pull/39 and allows us to have a more consistent diff whilst running `rules prepare`