Add new binary for measuring end to end alerting latency - e2ealerting (https://github.com/grafana/cortex-tools/pull/110)
* Add new binary e2ealerting

Helps measure the end to end alerting latency between
[Prometheus](https://github.com/prometheus/prometheus) scrapes and
[Alertmanager](https://github.com/prometheus/alertmanager) notification
delivery. We achieve that by creating a metric that exposes the current
time as its value, then creating an alert definition that uses that
value as an annotation, and finally a built-in
[Alertmanager](https://github.com/prometheus/alertmanager) webhook
receiver observes the duration between the value of the annotation and
webhook reception time as part of a Histogram.