alertmanager: Avoid breaking /api/v1/alerts if legacy prefix is empty (#3905)
If the legacy prefix is an empty string, it's interpreted as a handler for `/*`. Any handlers added afterwards are silently ignored, so requests to `/api/v1/alerts` are routed to the legacy handler, rather than to `am.*UserConfig` as would be expected. Meanwhile any handlers added BEFORE the `/*` legacy handler still work as expected - making this very difficult to debug!

Signed-off-by: Nick Parker <nick@nickbp.com>
Former-commit-id: a0c89bdf98e24ebfa4bf52a19cbd0bdc61a7b212