Increase CortexIngesterReachingSeriesLimit warning `for` duration
As it turns out, during normal shuffle-sharding operation, the 70%
mark is often exceeded, but not by much. Rather than increasing the
threshold to 75%, this commit increases the `for` duration to 3h,
following the thought that we want this alert to fire if ingesters are
constantly above the threshold even after stale series are flushed
(which occurs every 2h, when the TSDB head is compacted). We flush
series with a timestamp between [-3h, -1h] after the last compaction,
so the worst case scenario is that it takes 3h to flush a stale
series.

Signed-off-by: beorn7 <beorn@grafana.com>