Attempt fix to TestSingleBinaryWithMemberlistScaling.
The original fix was to check `memberlist_client_cluster_members_count`
to only scale down the next instance once the previous has been removed
everywhere. However, this does not inhibit the test enough to ensure
that tombstones are not lost, because the memberlist membership
propagates much quicker than ring membership. Therefore, instead wait
until every instance has seen the tombstone before removing another
instance.

This unfortunately takes the teeth out the test so the alternative is
to skip it or remove the test entirely, until issue #44 is fixed.