Replace uses of service.StopAsync in scheduler tests (#188)
Replace uses of the method service.StopAsync in scheduler tests that
expect the scheduler to return a SHUTTING_DOWN response. The new method
stops the service (scheduler) and waits until it has entered the
terminated state. This ensures that the tests don't fail if they run too
fast and the service has not completed terminating yet (and thus wouldn't
return a SHUTTING_DOWN response).

Fixes #186

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>