Increase allowed delta between sharded queriers in integration tests
Increase the allowed difference in number of queries handled by each sharded
querier during integration tests to %50 of the total number of queries. This
number is picked to verify that each querier handles _some_ non-trivial portion
of the requests but we allow for the variability we often see in these tests.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>