Use one background context for all ingester pushes
Create one context outside of `DoBatch()`, and use it for all ingesters.
We don't need a slightly different deadline for each; the timeout should
be orders of magnitude bigger than the time to start each call.

Move the `cancel()` call to the `DoBatch` cleanup function; this is
called after all workers finish, which is what we want.