Alertmanager: Prevent fallback configuration replacing existing config. (#224)
* Alertmanager: Prevent fallback configuration replacing existing config.

This attempts to solve a situation where blank (fallback) configurations can
overwrite a user configuration. When a request is made to an alertmanager
instance, we check if the user is active in the instance. If it is not, then a
blank configuration us uploaded and the user is activated in that instance.
There are two possible scenarios where this can lead to us overwriting an
existing user configuration:

1. If a request is served in the window of time between the configuration being
    written to object storage, and the configuration being read during the
    periodic polling. Due to the alertmanager instance not having that user
    active in memory yet, the fallback configuration can be uploaded.
2. When sharding is enabled, upon a user being moved between instances due to
    e.g. a scale up or down, then a period exists where the distributor may
    still be sending requests to an instance which is being deactivated.

To fix this, before uploading a blank configuration, simply check if one
already exists. This should eliminate the risk entirely during scale up/down
of fallback configurations being created for existing users.

This is still susceptible to a race condition of between the check/upload for
the first request ever made for a user. If the user uploads a configuration in
this window then the configuration is lost. However, this is the only window
of time.

An alternative to this fix was considered; uploading fallback configurations to
a different path, and checking that path in addition to the regular path when
reading configurations. This adds a lot of complexity however for (arguably)
very little gain.

As a secondary defensive measure, specifically for the sharding case which is
far more likely to happen, then we now check that the user is actually meant
to be present in that instance at the time of the request.

* Changelog.

* Return 406 when explicitly decided not to use fallback configuration.

* Review comments.

(cherry picked from commit 2b435f2e1b1be069a30292f48d1f00288b265b13)