Alertmanager: Use distributor for non-API routes. (#213)
* Alertmanager: Use distributor for non-API routes.

The web user interface routes were not passed through the distributor. This was
an oversight and likely worked until now because the user interface is purely
static content, and does not need to be served by the per-tenant instance. In
order to do this, instances with a fallback configuration were created as
needed to serve the content.

For consistency, all routes should go through the distributor, and this is
especially important if the distributor is ever run as a separate service.

* Changelog.

* Fix integration test.

* Only forward arbitrary GET/HEAD, not POST.

Co-authored-by: Marco Pracucci <marco@pracucci.com>
(cherry picked from commit 8e0eb3bad30d06444ea4751cadab566fdd88135a)