Correctly apply parallelism limits when doing querysharding.
Previously we were only splitting by day/time in the frontend, so applying the max parallelism was easy.
Now we can also apply parallelism at the querysharding level, and this means we can easily by-pass the `MaxQueryParallelism` limits.

This PR apply the limits at a lower level and so fixes the problem of overscheduling per query which we currently have since querysharding is activable.

This is inspired by work we've done in Loki.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>