Use concatenation for ShardSelector label value
While getting familiar with the code I saw this and I decided to do a
quick commit to replace a fmt.Printf call with a concatenation (and then
ended up using a stringbuilder)

This is 3x faster, well, we're talking about just 80ns, but still, why
waste time on parsing the format template, etc?
Usually the shard number will be small enough (<100) to be formatted
from a static array.

Also removed the fmt.Sprintf call from astmapper/sharding.go in favor of
ShardSelector{} receiver method call, to make that code cleaner, and
then moved the matcher instantiation to the receiver too so it's now
similar to the ShardSelector.Label() call.

All in all, this is the result of the benchmark comparison:

name \ time/op            fmt_printf.txt  concat_itoa.txt  stringbuilder.txt
ShardSelector_LabelValue      109ns ± 1%        38ns ± 1%          27ns ± 2%

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>