Set a default for ingester.instance-limits.max-inflight-push-requests
Set the default value for the ingester instance level setting of max inflight
push requests. This value is picked to be high enough that we shouldn't run into
it except in cases where the excessive number of requests would cause a spike
in request latency.

Setting this limit allows the ingester to reject sudden overwhelming bursts of
traffic instead of taking on more requests which cause memory usage to increase,
eventually leading to an OOM kill.

Fixes #248

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>