HA-Tracker: do KV-Store updates in the background
We don't want to hold up normal flow of push requests while we do a
compare-and-swap on KVStore, which can take hundreds of milliseconds.

Instead, store the last-seen timestamp of both the elected replica and
any other replica we see (presumably its fail-over pair), then update
KVStore periodically in the background.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>