Distributor tests: ensure ring goroutine is stopped
The Async version doesn't seem to do the job, because I get hundreds of
them running when I run the full test suite.

As the comment said, waiting for it to exit was slow, so speed this up
by waiting 40ms instead of 1 second to react.

It's still a bit slow so reduce the number of combinations of query
tests.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>