ha-tracker: avoid race with kv callback on cache update
Move the cache update out of `updateKVStore()`, so in general we will
only update the cache via a callback from the kv store.
One exception is if we are adding the very first record for a cluster,
as part of a push operation. Now we lock around that operation so we
don't overwrite an update from the kv callback.