indexed+diff+varint+snappy postings encoding
Profiles have shown that a lot of time of ExpandPostings is spent
decoding diff+varint+snappy postings, to be more specific, time is spent
doing the varint decoding. In the current diff+varint encoding we have
to decode all previous values in order to Seek() to a given position.

This is a new codec, that encodes into diff+varint and keeps an index
with page of given size (const defined, like 256), storing the offset of
each page's first value. So when calling Seek() we can skip entire
pages if the next pages's first value is still lower than the one we're
seeking for.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>