Add api/error package to return properly formatted errors (#360)
* Add JSONErrorf to return properly formated errors

All queryrange errors are currently not formated in a way that
Prometheus is formatting its errors.

* Map promql evaluation errors correctly

* Marshal JSON only when returning the error

This change moves the step of marshaling the error into the method when
its actually returned to the user.

It also determines the status code based on the errorType and renames
the API error method into Newf/New.

* Add faillint rule to ensure APIErrors are used by queryrange

* Add CHANGELOG entry

* Add provenance reference to prometheus

* Use correct error type for too long queries

* Rename GetRequest and PostRequest to DoGet DoPost

* Fix test for timeout error

(cherry picked from commit 1f4561938cdd185ffec75adb62305bdee5dc47b2)