SkipLabelNameValidation based on flag and header
Relates to https://github.com/grafana/mimir/issues/389

Before skipping label name validation on the http write path was not
possible. With this changes we've enabled a way to override this
behaviour using both the value of
`mimirpb.WriteRequest.SkipLabelNameValidation` and the value of the
request header `X-Mimir-AllowSkipLabelNameValidation`.

This will be opt in and only enabled via the api configuration flag
`api.allow-skip-label-name-validation-header-enabled`.

This change also updates unitary tests for both API and util/push
packages.

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>