Change ruler API to use HTTP 200 for list endpoint with no rules
Return an HTTP 200 and empty map from the `/api/v1/rules` endpoint
when there are no rules instead of an HTTP 404 since this does not
represent a client error.

Existing clients (cortextool and Grafana unified alerting) currently
don't handle the 404 well (cortextool) or work around the 404 and would
work correctly with a 200 (Grafana unified alerting).

Fixes #434

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>