SkipLabelNameValidation based on flag and header (#390)
* SkipLabelNameValidation based on flag and header

Relates to https://github.com/grafana/mimir/issues/389

Before skipping label name validation on the http write path was not
possible. With this changes we've enabled a way to override this
behaviour using both the value of
`mimirpb.WriteRequest.SkipLabelNameValidation` and the value of the
request header `X-Mimir-AllowSkipLabelNameValidation`.

This will be opt in and only enabled via the api configuration flag
`api.allow-skip-label-name-validation-header-enabled`.

This change also updates unitary tests for both API and util/push
packages.

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>

* Reject push request if disabled and header present

In case the `api.allow-skip-label-name-validation-header-enabled` is
disabled but a push request with header
`X-Mimir-AllowSkipLabelNameValidation` is sent the request should be
rejected with an http bad response error status code.

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>

* Update documentation

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>

* Update changelog

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>

* remove bufferpool.put in error handling

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>

* Accept PR feedback suggestions

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>

* Address pull request feedback

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>

* fix SkipLabelNameValidationHeader yaml attribute name

* Document skip label name validation in remote write

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>

* Address PR feedback and nitpicks

* move changelog entry to mimir / unreleased

Co-authored-by: Marco Pracucci <marco@pracucci.com>