Refactor NewMemcachedClient constructor
Previously NewMemcachedClient was returning the MemcachedClient
interface that only had a subset of cache methods. This was
limiting the use of the constructor to a single type of use.

The introduced change by this commit renames the constructor to
NewUpdatedMemcachedClient and refactors the return type that replaces
the interface with a specific type.

This enables the the constructor to be used for other purposes where the
caller may want to use other available methods in the memcached client
like Add, Delete, CompareAndSwap etcetera.