Rename MemcacheClient to GetterSetterMemcached
The interface was called MemcachedClient which conflicts with the actual
type that wraps the memcached client. So the change of this commit is to
rename the MemcachedClient to GetterSetterMemcached which gives more
information about what the interface actually does.

In a follow up commit we'll rename the UpdateMemcachedClient type to
simply MemcachedClient.