Stop embedding memcache client and actually implement methods
Before the refactor we did in the last two commits, we were only
exposing two memcache methods which were Set() and GetMulti(). With the
refactor and because MemcachedClient{} had a memcache client embedded we
were exposing any available methods in the client.

To make the behaviour consistent instead of embedding the memcache
client we're now defining the methods that we really want to expose.
Also this allows us to extend the behaviour like in the Set() method
where we're actually limiting the max item set or error wrapping where
needed.