Implement new methods for MemcachedClient
In order to provide a more complete MemcachedClient I've implemented
Add(), CompareAndSwap(), Get() and Delete(). This will remain unused in
the exisitng codebase but are available to offer a more complete public
API for any possible caller.

Also all the methods above are consistently limiting the max item size
and error wrapping the responses in case of error.