Mutex hat for the circuitbreaking map list
The MemcachedClient struct had a mutex defined at the top that was only
used for handling concurrent writes and reads to the cirbuit breaker
map. We've decided to place the mutex next to this map in the struct so
that its easier for a reader to spot the relationship between the two
variables.