Benchtool: add `-bench.write.proxy-url` argument. (https://github.com/grafana/cortex-tools/pull/223)
This allows configuring the Prometheus remote-write client with a HTTP proxy URL.

Getting the client to support the HTTP[S]_PROXY environment variables is not
straightforward, as the creation of the HTTP transport is deep in the upstream
code. Additionally, the client appears to have it's own mechanism for
configuring a HTTP proxy, so it would be safer to use that without meddling.