Add limit on label names and values time range
Flag is named `-store.max-labels-query-length`.
If set, this acts to restrict the range of label-names and label-values
requests, e.g. if you set to 30 days and request 90 days, you get the
result for the last 30 days of the range.
This is similar to the behaviour of `querier.max-query-lookback`.

(cherry picked from commit 90862edd0dc3a810368128cd39b60216f6db40c8)