Remove chunk storage engine from ingesters
This removes the deprecated chunks storage from the ingester component.
The use of chunks storage has been in maintenance mode for a while and
the first release of Mimir is supposed to release without chunks storage
engine.

This will be followed by more PR removing chunks related code and
functionality.

The flags removed as part of this commit are:
- `-ingester.checkpoint-duration`
- `-ingester.checkpoint-enabled`
- `-ingester.flush-on-shutdown-with-wal-enabled`
- `-ingester.recover-from-wal`
- `-ingester.wal-dir`
- `-ingester.wal-enabled`