Move /series max-range from blocks-store to querier
So that we can apply it before the `-store.max-query-length` limit is
checked, otherwise long `/series` requests just error.

Note `distributorQuerier.Select()` is now using `SelectHints` instead of
querier min/max time, so we need to move the ingester-only case
in `querier.Select()` after the time restrictions.

In turn this means we cannot pass `nil` for `SelectHints` into
`distributorQuerier.Select()`, which we used to do in a test.