Add limit on /series, label names and values time range (#507)
Flag is named `-store.max-labels-query-length`.

If set, this acts to restrict the range of `/series`, label-names and
label-values requests, e.g. if you set to 30 days and request 90 days,
you get the result for the last 30 days of the range. This is similar
to the behaviour of `querier.max-query-lookback`.

For label-name and label-values requests it is implemented on
`blocksStoreQuerier` rather than `querier`, because the latter has
no access to manipulate the time-range of the former.

Also some tidying up and additional trace logging.

* Fix querier select comment

We do limit the time-range of series and labels, in
`querier.NewQueryable`, so the comment was incorrect.

* refactor: extract func from validateQueryTimeRange

Save repeating ourselves.

* Tidy up handling of series requests

I think this version is clearer.

Also, previously if we did receive a nil hints and querying store was on, it would crash.

Note `distributorQuerier.Select()` is now using `SelectHints` instead of
querier min/max time, so we need to move the ingester-only case
in `querier.Select()` after the time restrictions.

In turn this means we cannot pass `nil` for `SelectHints` into
`distributorQuerier.Select()`, which we used to do in a test.

(cherry picked from commit 8b6c5e2afe4f78b9c3fca3805fc6658aae21fde4)