Querier: labels requests obey query-ingesters-within
If `-querier.query-ingesters-within` is set, make series, labels and
label-values requests to ingesters only within that time.

Ingesters may have data going back much further, depending on retention
setting, so it's more efficient to query just the window we won't get
from blocks storage.

Historically we only queried ingesters, so it was good to make the
request as long as possible, but now we always fetch labels from block
storage we don't need to.