Store: retrieve LabelValues intersecting Postings
Instead of fetching series, we can just intersect postings for each one
of the label values. This is especially performant for namespaces where
series are organised as a tree and reuse few label names.

This implementation doesn't take into account the req.Start and req.End
anymore, but that's also consistent with current implementation when no
matchers are specified, and nobody seems to complain about it.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>