Use singleflight for expanded postings promises
Replaces custom promise implementation in ExpandedPostings by
singleflight.Group implementation.

This didn't really result in less code, although most of the code now is
just boilerplate.

The bad thing about this is that we're now creating a goroutine here,
othwerwise we are unable to abandon the result on context cancellation
or timeout.

The good thing is that singleflight provides a `shared` flag, which we
can use to avoid having to copy the postings in the most common case (I
saw a lot of traces and very few of them actually reuse the promise,
most just hit the cache)

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>