Move cacheKey to inmemory.go and strings to MC
While writing a comment on cacheKey interface I realised that the
implementations had to be hasheable, but that wasn't something that
memcache needed, so looking deeper I came to the conclusion that this
unexported interface was adding too much coupling between Memcached and
Inmemory implementations, without too much benefit (apart from carrying
the type in the unexported calls of Memcached implementations).

So, I moved the string() method into separate functions used by the
memcached implementation, and the cacheKey interface to the inmemory
implementation.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>