Remove --quiet on some git diff's
Failures like this one https://github.com/grafana/mimir/runs/4507253663?check_suite_focus=true
are difficult to trace back to the file that caused them. This is
usually not a problem if the remediating `make` command actually works.
But in the case above, the GH actions workflow is probably working on a
stale commit. Being able to view the diff that's causing the workflow to
 fail will help to diagnose this.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>