storagegateway/gateway: wait for ring tokens stability instead of ring stability
During startup every storage gateway does this sequence in parallel:
set JOINING state in the ring, wait for ring to stabilize, set ACTIVE
state. Setting ACTIVE state violates the stability so some instances will
restart waiting on stability and the sequence starts over. This can happen
multiple times, e.g. in the test with 20 gateways, some just timed out after
30 seconds.

The observation is that we do not actually care if an instance is in JOINING
or ACTIVE state, we just want a stable Ring token distribution. The new
function in dskit is targetting this requirement. Startup should be quicker
and stability tests take a minute less.

See discussion in dskit PR 95

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>