Ingesters: respect start/end time on labels&series requests
Previously ingester would return all current data if you asked for an
old range. Now it will return data in range.

With the new behaviour, if you don't set `EndTimestampMs` then it
is treated as zero and you will get nothing back. This is on the
internal gRPC API and it different to the Prometheus API which treats no
end time as "till the end of time".

One test needed fixing as it actually requested data from the wrong
time - the zero value `time.Time{}` is a specific point in the past.
Make it request data for all time instead.

Fix up other tests to match the new behaviour.