Fix flaky TestIngester_inflightPushRequests
My machine is so fast that it can process more than 3600_000 samples in
a second, which means that this test tries to send more than 3600_000
samples, and with 1 millisecond per sample that is more than an hour of
samples so tsdb fails with an "out of bounds" error.

This trades samples for series, so we never send more than an hour of
samples.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>