Remove active series metrics when == 0
Similar to what was done with active trackers, this removes the active
series metric for a tenant as soon as it becomes zero.

This metric would be removed once (if) the tsdb becomes idle anyway, we
just do it earlier. This way we avoid all the raciness between the tsdb
closing and updating active metrics (as long as the tsdb idle timeout is
higher enough than the timeout to consider metrics inactive)

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>