Remove -distributor.shard-by-all-labels
Remove the `-distributor.shard-by-all-labels` configuration flag
and always assume it is true. This is a potentially disruptive change,
since this flag was not deprecated and previously defaulted to false.

A quote from the Cortex documentation about this flag's default behavior:
> In hindsight, this seems like the wrong choice: we do many orders of magnitude
> more writes than reads, and ingester reads are in-memory and cheap. It seems
> the right thing to do is to use all the labels to shard, improving load
> balancing and support for very high cardinality metrics.

Removing this also flag provides an opportunity to remove local limits, which
could help simplify configuration further.