Increase scheduler worker cancellation chan cap
With previous implementation, if worker was busy talking to scheduler,
we didn't push the cancellation, keeping that query running.

When cancelling a query, all its subqueries are cancelled at the same
time, so this was most likely happening all the time (first subquery
scheduled on this worker was canceled, the rest were not because worker
was busy cancelling the first one).

Also removed the `<-ctx.Done()` escape point when waiting for the
enqueueing ACK and modified the enqueueing method to ensure that it
always responds something.

Fixes: https://github.com/grafana/mimir/issues/740
Inspired by: https://github.com/grafana/loki/pull/5113

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>