Handle cancellations before requests
When scheduler worker is handling requests, if there's certain pressure
on the loop, we can keep enqueueing requests instead of handling the
cancellations. However, when there's that pressure on the loop, it's
more important than never to properly cancel the requests.

This checks the cancellations channel first, and then checks the
requests channel too.

Related to https://github.com/grafana/mimir/pull/741

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>