Add "user" label to appended exemplars metric
Currently there are three metrics on the received exemplars:
- distributed_in: total exemplars received, before deduplication.
- distributor_received: same but without duplicated and rejected.
- ingester_appended: the ones that were actually appended to the
storage, if it's enabled for the user.

The first two have per-user labels, but the last one doesn't, so we can
know whether we're receiving exemplars, but we don't know whether
they're actually being appended (or they're being dropped because of
limits configuration).

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>