Remove legacy HTTP endpoints
The non-legacy `/api/v1/rules/**` endpoints remain as they are. The
current legacy endpoints have their prefixes changed: from
`LegacyHTTPPrefix` (set with the `http_prefix` YAML option) to
`PrometheusHTTPPrefix` (set with `api.prometheus_http_prefix`).

The reason for this decision is to keep the API more consistent in the
way it groups API endpoints. If were to only remove the legacy prefixes,
 then the ruler configuration API (`POST|GET /api/v1/rules/**`) will be
 under a different prefix from the prometheus-compatible
 API (`GET /prometheus/api/v1/(rules|alerts)`).

The legacy endpoints will be deleted.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>