Set new default rule group limits
Previously the default for both of these limits was 0, which disabled them.
It is usually recommended to set limits then gradually increase them as needeed,
so it would be nice to set some defaults

These changes:
- Default `-ruler.max_rules_per_rule_group` to 20 (previously 0).
- Default `-ruler.max_rule_groups_per_tenant` to 70 (previously 0).

These values were chosen as they were the ones ultimately recommended to customers
in the latest configuration audit.

For reference, the current publicly recommended GEM limits can be found here:
https://grafana.com/docs/metrics-enterprise/latest/setup/limits/#ruler-max-rules-per-rule-group

`-ruler_max_rules_per_rule_group` set to 20 is the same as that recommendation.
`-ruler_max_rule_groups_per_tenant` is not the same, and is 70 here instead of 10.

From my recollection, these values weren't scientifically chosen and were more of a best guess
at what a reasonable value would be that didn't impede users initially, but offered them some assurance.