Tenant federation implementation of exemplar query interfaces
Allow queries with multiple tenants against exemplar endpoints. Queries
for each tenant are split and run in parallel using the same logic as the
existing tenant federation query endpoints.

Similar to the existing federated endpoints, the label `__tenant_id__` is
added to the set of series labels for each exemplar to indicate which tenant
it was queried from.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>