Test backward-(in)compatibility
Since some default flags are set when the service is being start, we
need to pass the flag mapping function from the backward compatibility
definition down to the service starting.

I created an options-like scheme for services (currently supporting only
flag mappers, but we could also move other options there, like
WithImage() instead of passing in an empty string when we want the
default image.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>