Change pull request and issue templates
We're using bold text as headings, instead of headings.
Markdown has headings, so there's no need to use bold text, plus github
UI sometimes gets buggy about bold-text-headings because it thinks that
the wildcard is a list item, so when you press enter it creates another
wildcard-list-entry.

I propose to fix that.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>