Add parenthesis to sharded avg expression
When we're sharding an avg() we actually do `sum() / count()`, however
when the expression is involved in something that has a special label
mataching, like:

    foo * on(a, b) group_left(c) avg by(a, b, c) (bar)

We shard into:

    foo * on(a, b) group_left(c) sum by(a, b, c) / count by(a, b, c) (bar)

Which doesn't work, because the second binary operation also need label
matching.

This PR adds a parenthesis around the sharded average, so it's treated
as a single output when performing label matching.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>