Extract HTML templates to .gohtml files
Instead of having HTML in constants, extract them to .gohtml, properly
annotated with the type they expect to be rendering, and then embed them
using the `//go:embed` directive.

Also standarize the usage across projects, variable for the HTML of page
`foo` is called `fooPageHTML`, the template is `fooPageTemplate` and the
struct that is being rendered is called `fooPageContents`.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>