Sync both latest and next documentation
Pushes to main are published to the `next` directory and release
branches and tags are published to the a versioned directory.

The `Determine technical documentation version` step coerces the git
reference into a version string of the form `v.<major>.<minor>` which
matches the versioned documentation structure on the website.

Syncing both latest and next documentation facilitates backporting
documentation fixes to release branches in order to update latest
documentation.

Also introduces the `test` job to ensure broken documentation is not
synced the to the website.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>