Bump mimir prometheus to discard unkown chunks
This commit bumps the vendored version of mimir-prometheus to bring the
changes included in https://github.com/grafana/mimir-prometheus/pull/196

These changes affect the TSDB Head replay of memory mapped chunks
ensuring that if a memory mapped chunk is read but it has an unknown
encoding, it is discarded rather than returning a corruption error.

This change is previous step towards releasing out of order support,
where a new chunk encoding is being released. This previous step would
allow us to rollback between versions if we ever get to that point.

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>