Use a dedicated thread pool for creating indexheader.Readers
Remove use of a dedicated thread pool for store-gateway requests. Instead,
only use the thread pool for creation of new indexheader.Reader instances
since this is when most I/O on memory mapped files happens.

Using dedicated threads for entire requests introduced a new way for store-
gateways to fail when becoming overloaded. Since the threads in the pool are
a constrained resource, when slow requests monopolize them all other requests
are blocked waiting for a thread. By moving use of the thread pool to
creation of indexheader.Reader instances, we minimize the amount of time
that a thread must be used for while also making sure that we don't use
the thread pool at all when not needed.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>