Helm: enable multi-tenancy in OSS by default
Disabled multi tenancy makes it look like OSS does not support it.
Also it makes it harder to migrate to multi tenancy and/or GEM.

Enable multi tenancy by default, but at the same time make sure that
Nginx sets the X-Scope-OrgID to anonymous if it's missing. This is done
for backwards compatibility and ease of use.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>