Update the bytes length of the meta files
With the new vendored version of mimir-prometheus the meta files now
have a new property called out_of_order to state if the block was an
out-of-order block or not.

This new property gets added to every meta file so we need to update the
expected lengths of this tests.

This is how the new property looks in the metas:
- "out_of_order": false,