Use minified sdk version for mimirtool
github.com/grafana-tools/sdk is a repository that is not maintained
by Grafana, and it can't properly parse all the dashboards
from Grafana because the model has evolved over time.

I tried to fix that in github.com/colega/grafana-tools-sdk but it
doesn't seem to be feasible to fix all and support all Grafana
dashboards.

Instead of that, we can just use our own model for the dashboards,
which is a stripped down version of github.com/grafana-tools/sdk with
only the fields that we need. This decreases the possibilities of
failing when parsing, while still does all the work.

With this, we can now drop the replace directive for my fork.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>