Implement tenant federation for metadata querier endpoint (#2467)
Allow multiple tenant IDs to be passed to the `/api/v1/metadata`
endpoint in the same way we support for query and exemplar endpoints.

For other endpoints, the tenant ID is included as a generated label
(query results or exemplars). We don't have the option here because
metadata doesn't include any labels. Because of this we simply merge
and de-duplicate metadata for all tenants. The `http.Handler` will group
all metadata under each metric name.

Fixes #2410

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>