Don't implement UnmarshalYAML on Mimir's config
Instead of having to wrap it with a type that removes that method, just
add a type that adds this method only when needed.

Whoever is using this config will have to explicitly call the
UnmarshalYAML.

This way we don't need to convert the type back every time we really
want to call some mimir.Config's method (like RegisterFlags).

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>