e2e tests: json-encoded MIMIR_EXTRA_FLAGS
This replaces the old MIMIR_EXTRA_ARGS which was a simple string that we
added to the args provided to the image.

That approach didn't solve all use cases, since it was appending the
flags after they've been mapped by the flagmapper, which means that same
flags were appended for all image versions, making this incompatible
with backwards compatibility tests.

Now we parse the MIMIR_EXTRA_FLAGS as JSON, merge it with all flags we
have, and then we apply the flagmapper, if any, properly mapping the
flags in case of backwards compatibility tests.

Also renamed _ARGS to _FLAGS since even though both are correct, in our
code we call them `flags`.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>