Helm-jsonnet: do not fail commenting for community PRs
It is not possible to automatically comment on external (to our public
mimir repo) PRs. As per documentation:

https://github.com/marketplace/actions/create-or-update-comment :
Note: In public repositories this action does not work in pull_request
workflows when triggered by forks. Any attempt will be met with the error,
Resource not accessible by integration. This is due to token restrictions
put in place by GitHub Actions.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>