Allow "all" clusters in dashboard templating vars
I understand why we don't want to select all namespaces in the
variables, as that would query too many metrics, bringing a useless
amount of series to the frontend.

However, allowing "All" on the "cluster" variable, allows selecting the
namespaces without having to select the cluster (which is often
redundant), while it still requires selecting one specific namespace.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>