Improve QueriesStuck alert by checking range min
When scraping metrics more than once per minute, the queue may have been
empty in some intermediate steps, but this alert would still be
triggered when the latest point is non-zero (for all the evaluations
in the 5m range).

This is a quick solution to that without investing too many resources
in developing a more sophisticated metric: we'll take a min_over_time of
that last minute every time, and we won't alert if it went to 0 at some
point in that minute (which is what we want).

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>