Dashboards: Introduce support for baremetal deployment
The following commit introduce the support of baremetal deployment for
*resources* dashboards. It's done by completing the hash
`resources_panel_series` and `resources_panel_queries`. The promql
queries are based on node_exporter.

To accommodate panels where the number of query differ we rely on two
intermediate functions `resourcesPanelLegend` and
`resourcesPanelQueries` who are responsible for returning the proper
number of queries/legends.

Signed-off-by: Wilfried Roset <wilfriedroset@users.noreply.github.com>